+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add Svelte syntax highlighting for code blocks (thanks to Axios team for the tip here!)
 - Adds example writable/derived store setup for dynamic information display
+- Fixes package name in documentation
 
 ## v0.7.1