+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Next
 
+## v0.7.2
+
 - Add Svelte syntax highlighting for code blocks (thanks to Axios team for the tip here!)
 - Adds example writable/derived store setup for dynamic information display
 - Fixes package name in documentation